requirements.txt
torch
torchvision
transformers
pillow
numpy
scikit-learn
tqdm
fastapi
uvicorn
python-multipart
undetected_chromedriver
bs4
requests